Update Dockerfile
eternnoir/activiti did not work out-of-the-box for me after a number of trials.  After cloning the repository, build failed apparently because it could not resolve 'dockerfile/java' repository which appears to have been replaced by 'java' repository so decided to use it instead.  The image build succeeded after the update.